requirements.txt
asgiref
Django
djangorestframework
sqlparse
Pillow
psycopg2-binary
python-dotenv